+++ examples/web-chat/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @smooai/smooth-operator-web-chat-example
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [b910a11]
+- Updated dependencies [c64b97b]
+  - @smooai/smooth-operator@1.27.4
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [0e65c59]
+  - @smooai/smooth-operator@1.27.3
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [e5bb69c]
+  - @smooai/smooth-operator@1.27.2
+
 ## 0.0.20
 
 ### Patch Changes